Auditorium refactor: move user types to types/User.ts
