+++ src/types/auditorium.ts
@@ -1,4 +1,4 @@
-import { DisplayUser } from "types/chat";
+import { DisplayUser } from "types/User";
 
 import { WithId } from "utils/id";
