+++ src/utils/selectors.ts
@@ -6,7 +6,6 @@ import { ArtCar, Firebarrel } from "types/animateMap";
 import { AuditoriumSection } from "types/auditorium";
 import {
   ChatSettings,
-  DisplayUser,
   JukeboxMessage,
   PrivateChatMessage,
   VenueChatMessage,
@@ -16,7 +15,7 @@ import { Reaction, TextReaction, TextReactionType } from "types/reactions";
 import { ScreeningRoomVideo } from "types/screeningRoom";
 import { Settings } from "types/settings";
 import { SparkleSelector } from "types/SparkleSelector";
-import { User, UserWithLocation } from "types/User";
+import { DisplayUser, User, UserWithLocation } from "types/User";
 import { AnyVenue, PosterPageVenue, VenueEvent } from "types/venues";
 
 import { WithId } from "utils/id";
